requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy<2
requests
python-dotenv

